team-editing: update team editing
